+++ packages/terra-base/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 5.40.0 - (September 29, 2020)
+
+* Added
+  * Added `throwOnI18nLoadError` prop to opt into logging and throwing the error when i18n data fails to load instead of only logging the error to the console
+
 ## 5.39.0 - (August 4, 2020)
 
 * Changed